Add files via upload
These four html files were produced from R markdown notebooks.  The Project Summary contains an overview of the case study. The Data Preparation and Analysis notebooks show the data manipulation, exploration, and analysis that was done. The Summary of Findings in the final product, with key findings, data visualizations, and recommendations.